+++ CHANGELOG.md
@@ -9,10 +9,15 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Features
 - adds version-based index mapping update support to the Search Relevance plugin [#344](https://github.com/opensearch-project/search-relevance/pull/344)
 - LLM Judgement Customized Prompt Template Implementation  [#264](https://github.com/opensearch-project/search-relevance/pull/264)
+- Add `_search` endpoint for searching for Search Configurations using OpenSearch DSL  [#372](https://github.com/opensearch-project/search-relevance/pull/372)
+- Add `_search` endpoint for searching for Judgments using OpenSearch DSL  [#371](https://github.com/opensearch-project/search-relevance/pull/371)
 
 ### Enhancements
 - Add BWC and Integration Tests for Index Mapping Update with Schema Version [#349](https://github.com/opensearch-project/search-relevance/pull/349)
 - Added better version of ESCI demo dataset that has images and overlaps with our ESCI judgment data.  More compelling demonstrations.  ([#354](https://github.com/opensearch-project/search-relevance/pull/354))
+- Add supports to parse custom UBI indexes  [#364](https://github.com/opensearch-project/search-relevance/pull/364)
+
+- Add `/experiments/_search` endpoint to allow searching for experiments using OpenSearch query DSL ([#369](https://github.com/opensearch-project/search-relevance/pull/369))
 - Support for adding description in Search Configuration ([#370](https://github.com/opensearch-project/search-relevance/pull/370))
 
 ### Bug Fixes
@@ -24,5 +29,6 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Documentation
 
 ### Maintenance
+- Fix jackson annotations version ([#374](https://github.com/opensearch-project/search-relevance/pull/374))
 
 ### Refactoring